Move some dependencies from direct to peer
After the latest round of ecosystem shenanigans
(https://github.com/dominictarr/event-stream/issues/116) I want to
reduce the amount of direct dependencies we employ, and as a result
the number of packages we have to put our trust in. By moving
dependencies to peer dependencies where it makes sense, we greatly
reduce the number of packages that we're directly responsible for
trusting.