Remove `Err#get` and `Ok#getErr`
They have been replaced with `Result#getUnsafe` and all call sites have
been updated.

It was discovered that `Request#from` and `Response#from` were unsafely
unwrapping the result of parsing an URL that could be invalid which
would result in an exception being thrown. It was decided to instead
propagate the error up to the caller which means introducinga breaking
change to the public interface. Users can choose to call `.getUnsafe()`
to retain the behavior or they can also choose to handle the result.