Add `Result#getUnsafe` (#1395)
* Remove `Err#get` and `Ok#getErr`

They have been replaced with `Result#getUnsafe` and all call sites have
been updated.

It was discovered that `Request#from` and `Response#from` were unsafely
unwrapping the result of parsing an URL that could be invalid which
would result in an exception being thrown. It was decided to instead
propagate the error up to the caller which means introducinga breaking
change to the public interface. Users can choose to call `.getUnsafe()`
to retain the behavior or they can also choose to handle the result.

* Add changeset

* Update changeset

* Update documentation

* Update changeset

* Update tests to only check that return value is Err

* Update changeset

* Update test that was overlooked in previous commit

* Implement review suggestions

* Change word in changeset

* Change wording in other changeset

* Remove unnecessary comment

* Update api documentation

* Remove another useless comment

* Update .changeset/rotten-walls-attack.md

Co-authored-by: Jean-Yves Moyen <jym@siteimprove.com>

* Update packages/alfa-style/src/property/box-shadow.ts

Co-authored-by: Jean-Yves Moyen <jym@siteimprove.com>

---------

Co-authored-by: Jean-Yves Moyen <jym@siteimprove.com>